+++ src/components/Table/TableTd.vue
@@ -20,11 +20,11 @@ const props = withDefaults(defineProps<ITableTd>(), {
 
 <style>
 .table-td {
-  padding: var(--td-vertical-padding) var(--td-horizontal-padding);
+  padding: var(--td-vertical-padding, 0) var(--td-horizontal-padding, 0);
 
   border: none;
-  border-left: 1px solid var(--border-grey);
-  border-top: 1px solid var(--border-grey);
+  border-left: 0.5px solid var(--border-grey);
+  border-top: 0.5px solid var(--border-grey);
   font-size: var(--td-font-size, 14px);
 }